EES-3453 Changes in response to PR comments
