+++ src/explore-education-statistics-common/src/modules/release/components/ReleaseDataFilesAccordion.module.scss
@@ -1,8 +1,18 @@
 @import '~govuk-frontend/govuk/base';
 
-.additionalInfoContainer {
+.container {
   margin-bottom: govuk-spacing(9);
   margin-top: -(govuk-spacing(6));
+
+  @include govuk-media-query($until: tablet) {
+    margin-top: 0;
+  }
+
+  @include govuk-media-query($from: tablet) {
+    :global(.govuk-accordion__section-header) {
+      border-top: 0;
+    }
+  }
 }
 
 .section {
