+++ tests/robot-tests/tests/admin_and_public/bau/publish_release_and_amend.robot
@@ -251,8 +251,8 @@ Verify public pre-release access list
 
 Verify accordions are correct
     user goes to release page via breadcrumb    ${PUBLICATION_NAME}    ${RELEASE_NAME}
-    user checks there are x accordion sections    1    id:additional-info-accordion
-    user checks accordion is in position    Explore data and files    1    id:additional-info-accordion
+    user checks there are x accordion sections    1    id:data-accordion
+    user checks accordion is in position    Explore data and files    1    id:data-accordion
 
     user checks there are x accordion sections    2    id:content
     user checks accordion is in position    Dates data block    1    id:content
